update with bruce's comment & solve issues
