+++ libs/movie/src/lib/movie/form/promotional-elements/promotional-elements.component.html
@@ -62,4 +62,9 @@
       </div>
     </ng-container>
   </div>
+  <ng-container>	
+    <button mat-icon-button color="primary" (click)="promotionalElements.addPromotionalElement()">	
+      <mat-icon>add_circle</mat-icon>	
+    </button>	
+  </ng-container>
 </form>
